refac: 리뷰 반영 - Lottos 파일을 LottoTicket, LottoTickets로 분리
