+++ src/main/kotlin/lotto/domain/LottoTickets.kt
@@ -2,22 +2,6 @@ package lotto.domain
 
 import lotto.constant.WinningInfo
 
-data class LottoTicket(
-    val lottoNumbers: List<Int>
-) {
-    init {
-        require(lottoNumbers.size == LOTTO_SIZE) { "로또 번호는 6개로 구성되어야 해요" }
-    }
-    fun matchCount(winningNumber: WinningNumber): Int {
-        return lottoNumbers.count { lottoNumber ->
-            winningNumber.winningNumbers.contains(lottoNumber)
-        }
-    }
-    companion object {
-        const val LOTTO_SIZE = 6
-    }
-}
-
 data class LottoTickets(
     val lottoCount: Int,
     val lottoTickets: List<LottoTicket>
